Move adminLookupMetaStream to separate module
